+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Code cleanup, minor enhancements.
+
+### Removed
+- iOS: Removed the `FontSize` constants.
 
 ## [2.0.0-beta.6] - 2018-08-21